1.4_project_clean/requirements.txt
annotated-types==0.5.0
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
click==8.1.7
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
fonttools==4.42.1
greenlet==2.0.2
h11==0.14.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
matplotlib==3.8.0
matplotlib-inline==0.1.6
mysql-connector==2.2.9
mysql-connector-python==8.1.0
nest-asyncio==1.5.8
numpy==1.26.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.4.1
pydantic_core==2.10.1
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
seaborn==0.12.2
six==1.16.0
SQLAlchemy==2.0.21
stack-data==0.6.2
tornado==6.3.3
traitlets==5.10.1
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.23.2
wcwidth==0.2.6
wordcloud==1.9.2
